Removed collapsable <Project> panels, style changes, API query is now made within App.js
