+++ 6.9.4
@@ -1,0 +1,1 @@
+/usr/local/n/versions/node/6.9.4
