+++ package.json
@@ -7,11 +7,13 @@
     "body-parser": "^1.17.1",
     "emailjs": "^1.0.8",
     "morgan": "^1.8.1",
+    "nodemailer": "^4.0.1",
     "react": "^15.4.2",
     "react-bootstrap": "^0.30.8",
     "react-dom": "^15.4.2",
     "react-router": "4.0.0-alpha.3",
-    "react-router-dom": "^4.0.0"
+    "react-router-dom": "^4.0.0",
+    "yarn": "^0.24.5"
   },
   "devDependencies": {
     "babel-cli": "^6.24.0",
