--- src/package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "nodemailer": "^4.0.1"
-  }
-}
